Practice mode?
At this point and time, the kana sprites and the sound sprites are direct reflections of each other (i.e., the first kana sprite's sound matches the first kana sound sprite; meaning you can easily guess the sprite to the sound)

Also, both the kana sprites and the kana sounds are completely randomized; this setting will familiarize the player with all of the kana to the sound. It's like having a random deck of flashcards, but the flashcard shows both the term and the definition.